Continued progress on convolution2.rs. Still need to modify main.rs.

+++ src/convolutional.rs
@@ -20,7 +20,7 @@ pub fn kernelize_all(labeled_images: &Vec<(u8,Image)>, levels: usize) -> Vec<(u8
     kernelized
 }
 
-pub fn kernelized_distance(k1: &Vec<Image>, k2: &Vec<Image>) -> f64 {
+pub fn kernelized_distance(k1: &Vec<Image>, k2: &Vec<Image>) -> u32 {
     assert_eq!(k1.len(), k2.len());
     (0..k1.len()).map(|i| euclidean_distance(&k1[i], &k2[i])).sum()
 }
@@ -54,10 +54,10 @@ pub fn apply_kernel_to(img: &Image, kernel: &Image) -> Image {
     result
 }
 
-pub fn pixelize(distance: f64) -> u8 {
-    let max_distance = ((std::u8::MAX as f64).powf(2.0) * (KERNEL_SIZE.pow(2) as f64)).powf(0.5);
-    let distance_to_pixel_scale = (std::u8::MAX as f64) / max_distance;
-    (distance.powf(0.5) * distance_to_pixel_scale) as u8
+pub fn pixelize(distance: u32) -> u8 {
+    let max_distance = ((u8::MAX as f64).powf(2.0) * (KERNEL_SIZE.pow(2) as f64)).powf(0.5);
+    let distance_to_pixel_scale = (u8::MAX as f64) / max_distance;
+    ((distance as f64).powf(0.5) * distance_to_pixel_scale) as u8
 }
 
 pub fn add_kernels_from_to(img: &Image, raw_filters: &mut Vec<Image>, kernel_size: usize) {
